Japanese dates: https://github.com/metanorma/metanorma-jis/issues/12
