+++ metanorma-jis.gemspec
@@ -27,9 +27,10 @@ Gem::Specification.new do |spec|
     f.match(%r{^(test|spec|features|bin|.github)/}) \
     || f.match(%r{Rakefile|bin/rspec})
   end
-  spec.test_files    = `git ls-files -- {spec}/*`.split("\n")
+  spec.test_files = `git ls-files -- {spec}/*`.split("\n")
   spec.required_ruby_version = Gem::Requirement.new(">= 2.7.0")
 
+  spec.add_dependency "japanese_calendar", "~> 0"
   spec.add_dependency "metanorma-iso", "~> 2.6.0"
   spec.add_dependency "pubid-jis"
